splitting up some files, cleaning house, and fixing import statements
