+++ .gitignore
@@ -19,3 +19,4 @@ __pycache__/time_modulation.cpython-39.pyc
 TheAddictionYouDidntKnowYouHad.mp3
 .idea/misc.xml
 src/nohup.out
+src/config.py
